Provide a convenience function to interleave stderr with stdout
This patch adds a convenience function which allows to interleave stderr with
stdout output. A shared pipe to the child process is used to read stdout and stderr.

Main motivation to use such a function is for debugging process output. An
example is writing acceptance tests. In order to figure out when errors
correlate with a behaviour it is much easier to see both stdout and stderr
output in comparison to have two separate process outputs.

Fixes https://github.com/fpco/typed-process/issues/16